legacy: removal of inveniocfg command
* INCOMPATIBLE Removes deprecated `inveniocfg` command line interface.
